+++ lib/propane/library_loader.rb
@@ -2,8 +2,7 @@
 
 # The processing wrapper module
 module Propane
-  require_relative 'ruby_library'
-  require_relative 'java_library'
+  require_relative 'library'
 
   # Encapsulate library loader functionality as a class
   class LibraryLoader
@@ -35,15 +34,11 @@ module Propane
     def loader(name)
       return true if @loaded_libraries.include?(name)
       fname = name.to_s
-      if (@library = LocalRubyLibrary.new(fname)).exist?
-      elsif (@library = InstalledRubyLibrary.new(fname)).exist?
-        return require_library(library, name)
-      end
-      if (@library = LocalJavaLibrary.new(fname)).exist?
-      elsif (@library = InstalledJavaLibrary.new(fname)).exist?
-        return load_jars(library, name)
-      end
-      false
+      library = Library.new(fname)
+      library.locate
+      return require_library(library, name) if library.ruby?
+      return false if library.path.empty?
+      load_jars(library, name)
     end
 
     def load_jars(lib, name)
